% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss16.R
\docType{data}
\name{gss16}
\alias{gss16}
\title{2016 General Social Survey}
\format{An object of class \code{data.frame} with 2867 rows and 17 columns.}
\usage{
gss16
}
\description{
A data set containing data on work, salary, and education from the 2016 General Social Survey.
}
\keyword{datasets}
