% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss16_Miss.R
\docType{data}
\name{gss16_Miss}
\alias{gss16_Miss}
\title{2016 General Social Survey (Missing Data)}
\format{An object of class \code{data.frame} with 2867 rows and 17 columns.}
\usage{
gss16_Miss
}
\description{
A data set containing data on work, salary, and education from the 2016 General Social Survey.
Missing data are not explicitly identified with NAs.
}
\keyword{datasets}
