% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDriveR-package.R
\docType{package}
\name{testDriveR}
\alias{testDriveR}
\alias{testDriveR-package}
\title{testDriveR: Teaching Data for Statistics and Data Science}
\description{
The goal of \code{testDriveR} is to provide data sets for teaching
    statistics and data science courses. This package includes a sample of
    data from John Edmund Kerrich's famous coinflip experiment. These are data
    that I use for teaching \href{https://slu-soc5050.github.io}{SOC 4015 / SOC 5050}
    at \href{https://www.slu.edu}{Saint Louis University}.
}
\details{
There are currently five data sets that are included in the package:

    \itemize{
      \item \code{\link{auto17}} - A data set containing model year 2017 vehicles
          for sale in the United States
      \item \code{\link{childMortality}} - A data set containing childhood mortality
          time series data by country from UNICEF
      \item \code{\link{gss14}} - A data set containing a selection of variables
          related to work and education from the 2014 General Social Survey
      \item \code{\link{gss14_simple}} - A simple version of \code{gss14} without
          factors created and without missing data explicitly declared
      \item \code{\link{kerrich}} - A data set containing 2000 trials of coin flips by
          \href{https://en.wikipedia.org/wiki/John_Edmund_Kerrich}{John Edmund Kerrich}
    }
}
